cluster: extraction logical changes
drain_status
security/plugin/cluster-link rpc types
controller_utils
rpc_utils

Split four logically separable pieces out of oversized headers into
their own files:

  - drain_status, previously nested in cluster::drain_manager, is now a
    freestanding struct in cluster/drain_status.h so consumers can use
    it without pulling in the full drain_manager (and its transitive
    partition_manager dep).
  - security_types.h, plugin_rpc_types.h, cluster_link_rpc_types.h
    are carved out of the 30+ structs in types.h.
  - controller_utils.{h,cc} and rpc_utils.{h,cc} are carved out of
    cluster_utils.{h,cc}.

No functional change: all new files are currently built as part of the
//src/v/cluster:cluster monolith target, and the donor headers carry
bridging includes so every existing consumer keeps resolving the same
symbols transitively. A follow-up commit breaks :cluster into
fine-grained targets and updates consumers to take direct deps.

The one forced callsite change: cluster::drain_manager::drain_status is
renamed to cluster::drain_status because the nested struct no longer
exists on drain_manager.
