+++ src/v/cluster/BUILD
@@ -71,6 +71,7 @@ redpanda_cc_rpc_library(
     name = "tx_manager_migrator_rpc",
     src = "migrations/tx_manager_migrator.json",
     out = "tx_manager_migrator_service.h",
+    visibility = ["//visibility:public"],
 )
 
 redpanda_cc_rpc_library(
@@ -349,6 +350,7 @@ redpanda_cc_library(
         "controller_probe.cc",
         "controller_snapshot.cc",
         "controller_stm.cc",
+        "controller_utils.cc",
         "crash_reporter.cc",
         "data_migrated_resources.cc",
         "data_migration_backend.cc",
@@ -411,6 +413,7 @@ redpanda_cc_library(
         "rm_partition_frontend.cc",
         "rm_stm.cc",
         "rm_stm_types.cc",
+        "rpc_utils.cc",
         "scheduling/allocation_node.cc",
         "scheduling/allocation_state.cc",
         "scheduling/allocation_strategy.cc",
@@ -500,6 +503,7 @@ redpanda_cc_library(
         "cluster_link/fwd.h",
         "cluster_link/table.h",
         "cluster_link/types.h",
+        "cluster_link_rpc_types.h",
         "cluster_recovery_manager.h",
         "cluster_recovery_reconciler.h",
         "cluster_recovery_state.h",
@@ -518,6 +522,7 @@ redpanda_cc_library(
         "controller_service.h",
         "controller_snapshot.h",
         "controller_stm.h",
+        "controller_utils.h",
         "crash_reporter.h",
         "data_migrated_resources.h",
         "data_migration_backend.h",
@@ -533,6 +538,7 @@ redpanda_cc_library(
         "distributed_kv_stm.h",
         "distributed_kv_stm_types.h",
         "drain_manager.h",
+        "drain_status.h",
         "ephemeral_credential_frontend.h",
         "ephemeral_credential_rpc_service.h",
         "ephemeral_credential_serde.h",
@@ -595,6 +601,7 @@ redpanda_cc_library(
         "partition_recovery_manager.h",
         "plugin_backend.h",
         "plugin_frontend.h",
+        "plugin_rpc_types.h",
         "plugin_table.h",
         "prefix_truncate_record.h",
         "producer_state.h",
@@ -606,6 +613,7 @@ redpanda_cc_library(
         "rm_partition_frontend.h",
         "rm_stm.h",
         "rm_stm_types.h",
+        "rpc_utils.h",
         "scheduling/allocation_node.h",
         "scheduling/allocation_state.h",
         "scheduling/allocation_strategy.h",
@@ -621,6 +629,7 @@ redpanda_cc_library(
         "scheduling/types.h",
         "security_frontend.h",
         "security_manager.h",
+        "security_types.h",
         "self_test/cloudcheck.h",
         "self_test/diskcheck.h",
         "self_test/metrics.h",
@@ -791,6 +800,7 @@ redpanda_cc_library(
         "//src/v/utils:expiring_promise",
         "//src/v/utils:file_io",
         "//src/v/utils:fixed_string",
+        "//src/v/utils:functional",
         "//src/v/utils:hdr_hist",
         "//src/v/utils:human",
         "//src/v/utils:lazy_abort_source",
