+++ src/v/cluster/types.h
@@ -17,11 +17,14 @@
 #include "cloud_storage/remote_label.h"
 #include "cluster/cloud_metadata/cluster_manifest.h"
 #include "cluster/cluster_link/errc.h"
+#include "cluster/cluster_link_rpc_types.h"
 #include "cluster/errc.h"
 #include "cluster/feature_update_action.h"
 #include "cluster/fwd.h"
 #include "cluster/nt_revision.h"
 #include "cluster/partition_balancer_types.h"
+#include "cluster/plugin_rpc_types.h"
+#include "cluster/security_types.h"
 #include "cluster/topic_configuration.h"
 #include "cluster/tx_errc.h"
 #include "cluster/version.h"
@@ -1399,134 +1402,6 @@ struct topic_table_ntp_delta {
     fmt::iterator format_to(fmt::iterator it) const;
 };
 
-struct create_acls_cmd_data
-  : serde::envelope<
-      create_acls_cmd_data,
-      serde::version<0>,
-      serde::compat_version<0>> {
-    static constexpr int8_t current_version = 1;
-    std::vector<security::acl_binding> bindings;
-
-    friend bool operator==(
-      const create_acls_cmd_data&, const create_acls_cmd_data&) = default;
-    fmt::iterator format_to(fmt::iterator it) const {
-        return fmt::format_to(it, "{{ bindings: {} }}", bindings);
-    }
-
-    auto serde_fields() { return std::tie(bindings); }
-};
-
-struct create_acls_request
-  : serde::envelope<
-      create_acls_request,
-      serde::version<0>,
-      serde::compat_version<0>> {
-    create_acls_cmd_data data;
-    model::timeout_clock::duration timeout;
-
-    create_acls_request() noexcept = default;
-    create_acls_request(
-      create_acls_cmd_data data, model::timeout_clock::duration timeout)
-      : data(std::move(data))
-      , timeout(timeout) {}
-
-    friend bool operator==(
-      const create_acls_request&, const create_acls_request&) = default;
-    fmt::iterator format_to(fmt::iterator it) const {
-        return fmt::format_to(
-          it, "{{ data: {}, timeout: {} }}", data, timeout.count());
-    }
-
-    auto serde_fields() { return std::tie(data, timeout); }
-};
-
-struct create_acls_reply
-  : serde::
-      envelope<create_acls_reply, serde::version<0>, serde::compat_version<0>> {
-    std::vector<errc> results;
-
-    friend bool
-    operator==(const create_acls_reply&, const create_acls_reply&) = default;
-    fmt::iterator format_to(fmt::iterator it) const {
-        return fmt::format_to(it, "{{ results: {} }}", results);
-    }
-
-    auto serde_fields() { return std::tie(results); }
-};
-
-struct delete_acls_cmd_data
-  : serde::envelope<
-      delete_acls_cmd_data,
-      serde::version<0>,
-      serde::compat_version<0>> {
-    static constexpr int8_t current_version = 1;
-    std::vector<security::acl_binding_filter> filters;
-
-    friend bool operator==(
-      const delete_acls_cmd_data&, const delete_acls_cmd_data&) = default;
-    fmt::iterator format_to(fmt::iterator it) const {
-        return fmt::format_to(it, "{{ filters: {} }}", filters);
-    }
-
-    auto serde_fields() { return std::tie(filters); }
-};
-
-// result for a single filter
-struct delete_acls_result
-  : serde::envelope<
-      delete_acls_result,
-      serde::version<0>,
-      serde::compat_version<0>> {
-    errc error;
-    std::vector<security::acl_binding> bindings;
-
-    friend bool
-    operator==(const delete_acls_result&, const delete_acls_result&) = default;
-    fmt::iterator format_to(fmt::iterator it) const {
-        return fmt::format_to(
-          it, "{{ error: {} bindings: {} }}", error, bindings);
-    }
-
-    auto serde_fields() { return std::tie(error, bindings); }
-};
-
-struct delete_acls_request
-  : serde::envelope<
-      delete_acls_request,
-      serde::version<0>,
-      serde::compat_version<0>> {
-    delete_acls_cmd_data data;
-    model::timeout_clock::duration timeout;
-
-    delete_acls_request() noexcept = default;
-    delete_acls_request(
-      delete_acls_cmd_data data, model::timeout_clock::duration timeout)
-      : data(std::move(data))
-      , timeout(timeout) {}
-
-    friend bool operator==(
-      const delete_acls_request&, const delete_acls_request&) = default;
-    fmt::iterator format_to(fmt::iterator it) const {
-        return fmt::format_to(it, "{{ data: {} timeout: {} }}", data, timeout);
-    }
-
-    auto serde_fields() { return std::tie(data, timeout); }
-};
-
-struct delete_acls_reply
-  : serde::
-      envelope<delete_acls_reply, serde::version<0>, serde::compat_version<0>> {
-    std::vector<delete_acls_result> results;
-
-    friend bool
-    operator==(const delete_acls_reply&, const delete_acls_reply&) = default;
-    fmt::iterator format_to(fmt::iterator it) const {
-        return fmt::format_to(it, "{{ results: {} }}", results);
-    }
-
-    auto serde_fields() { return std::tie(results); }
-};
-
 using transfer_leadership_request = raft::transfer_leadership_request;
 using transfer_leadership_reply = raft::transfer_leadership_reply;
 
@@ -1772,27 +1647,6 @@ struct feature_update_license_update_cmd_data
     auto serde_fields() { return std::tie(redpanda_license); }
 
     fmt::iterator format_to(fmt::iterator it) const;
-};
-
-struct user_and_credential
-  : serde::envelope<
-      user_and_credential,
-      serde::version<0>,
-      serde::compat_version<0>> {
-    static constexpr int8_t current_version = 0;
-
-    user_and_credential() = default;
-    user_and_credential(
-      security::credential_user&& username_,
-      security::scram_credential&& credential_)
-      : username(std::move(username_))
-      , credential(std::move(credential_)) {}
-    friend bool operator==(
-      const user_and_credential&, const user_and_credential&) = default;
-    auto serde_fields() { return std::tie(username, credential); }
-
-    security::credential_user username;
-    security::scram_credential credential;
 };
 
 struct cluster_recovery_init_state
@@ -2928,29 +2782,6 @@ struct revert_cancel_partition_move_reply
       const revert_cancel_partition_move_reply&) = default;
 };
 
-struct upsert_role_cmd_data
-  : serde::
-      envelope<upsert_role_cmd_data, serde::version<0>, serde::version<0>> {
-    security::role_name name;
-    security::role role;
-
-    auto serde_fields() { return std::tie(name, role); }
-
-    friend bool operator==(
-      const upsert_role_cmd_data&, const upsert_role_cmd_data&) = default;
-};
-
-struct delete_role_cmd_data
-  : serde::
-      envelope<delete_role_cmd_data, serde::version<0>, serde::version<0>> {
-    security::role_name name;
-
-    auto serde_fields() { return std::tie(name); }
-
-    friend bool operator==(
-      const delete_role_cmd_data&, const delete_role_cmd_data&) = default;
-};
-
 /**
  * Broker state transitions are coordinated centrally as opposite to
  * configuration which change is requested by the described node itself. Broker
@@ -3174,65 +3005,6 @@ struct controller_committed_offset_reply
     auto serde_fields() { return std::tie(last_committed, result); }
 };
 
-/**
- * Create/update a (Wasm) plugin.
- */
-struct upsert_plugin_request
-  : serde::envelope<
-      upsert_plugin_request,
-      serde::version<0>,
-      serde::compat_version<0>> {
-    model::transform_metadata transform;
-    model::timeout_clock::duration timeout{};
-
-    friend bool operator==(
-      const upsert_plugin_request&, const upsert_plugin_request&) = default;
-
-    auto serde_fields() { return std::tie(transform, timeout); }
-};
-struct upsert_plugin_response
-  : serde::envelope<
-      upsert_plugin_response,
-      serde::version<0>,
-      serde::compat_version<0>> {
-    errc ec;
-
-    friend bool operator==(
-      const upsert_plugin_response&, const upsert_plugin_response&) = default;
-
-    auto serde_fields() { return std::tie(ec); }
-};
-
-/**
- * Remove a (Wasm) plugin.
- */
-struct remove_plugin_request
-  : serde::envelope<
-      remove_plugin_request,
-      serde::version<0>,
-      serde::compat_version<0>> {
-    model::transform_name name;
-    model::timeout_clock::duration timeout{};
-
-    friend bool operator==(
-      const remove_plugin_request&, const remove_plugin_request&) = default;
-
-    auto serde_fields() { return std::tie(name, timeout); }
-};
-struct remove_plugin_response
-  : serde::envelope<
-      remove_plugin_response,
-      serde::version<0>,
-      serde::compat_version<0>> {
-    uuid_t uuid;
-    errc ec;
-
-    friend bool operator==(
-      const remove_plugin_response&, const remove_plugin_response&) = default;
-
-    auto serde_fields() { return std::tie(uuid, ec); }
-};
-
 struct update_partition_replicas_cmd_data
   : serde::envelope<
       update_partition_replicas_cmd_data,
@@ -3338,273 +3110,6 @@ struct set_partition_shard_reply
       const set_partition_shard_reply&) = default;
 
     auto serde_fields() { return std::tie(ec); }
-};
-
-struct upsert_cluster_link_request
-  : serde::envelope<
-      upsert_cluster_link_request,
-      serde::version<0>,
-      serde::compat_version<0>> {
-    ::cluster_link::model::metadata metadata;
-    model::timeout_clock::duration timeout{};
-
-    friend bool operator==(
-      const upsert_cluster_link_request&,
-      const upsert_cluster_link_request&) = default;
-    auto serde_fields() { return std::tie(metadata, timeout); }
-};
-
-struct upsert_cluster_link_response
-  : serde::envelope<
-      upsert_cluster_link_response,
-      serde::version<0>,
-      serde::compat_version<0>> {
-    cluster::cluster_link::errc ec;
-
-    friend bool operator==(
-      const upsert_cluster_link_response&,
-      const upsert_cluster_link_response&) = default;
-    auto serde_fields() { return std::tie(ec); }
-};
-
-struct remove_cluster_link_request
-  : serde::envelope<
-      remove_cluster_link_request,
-      serde::version<0>,
-      serde::compat_version<0>> {
-    ::cluster_link::model::delete_shadow_link_cmd cmd;
-    model::timeout_clock::duration timeout{};
-
-    friend bool operator==(
-      const remove_cluster_link_request&,
-      const remove_cluster_link_request&) = default;
-
-    auto serde_fields() { return std::tie(cmd, timeout); }
-};
-
-struct remove_cluster_link_response
-  : serde::envelope<
-      remove_cluster_link_response,
-      serde::version<0>,
-      serde::compat_version<0>> {
-    cluster::cluster_link::errc ec;
-
-    friend bool operator==(
-      const remove_cluster_link_response&,
-      const remove_cluster_link_response&) = default;
-
-    auto serde_fields() { return std::tie(ec); }
-};
-
-struct add_mirror_topic_request
-  : serde::envelope<
-      add_mirror_topic_request,
-      serde::version<0>,
-      serde::compat_version<0>> {
-    ::cluster_link::model::id_t link_id;
-    ::cluster_link::model::add_mirror_topic_cmd cmd;
-    model::timeout_clock::duration timeout{};
-
-    friend bool
-    operator==(const add_mirror_topic_request&, const add_mirror_topic_request&)
-      = default;
-
-    auto serde_fields() { return std::tie(link_id, cmd, timeout); }
-};
-
-struct add_mirror_topic_response
-  : serde::envelope<
-      add_mirror_topic_response,
-      serde::version<0>,
-      serde::compat_version<0>> {
-    cluster_link::errc ec{cluster_link::errc::success};
-
-    friend bool operator==(
-      const add_mirror_topic_response&,
-      const add_mirror_topic_response&) = default;
-
-    auto serde_fields() { return std::tie(ec); }
-};
-
-struct update_mirror_topic_status_request
-  : serde::envelope<
-      update_mirror_topic_status_request,
-      serde::version<0>,
-      serde::compat_version<0>> {
-    ::cluster_link::model::id_t link_id;
-    ::cluster_link::model::update_mirror_topic_status_cmd cmd;
-    model::timeout_clock::duration timeout{};
-
-    friend bool operator==(
-      const update_mirror_topic_status_request&,
-      const update_mirror_topic_status_request&) = default;
-
-    auto serde_fields() { return std::tie(link_id, cmd, timeout); }
-};
-
-struct update_mirror_topic_status_response
-  : serde::envelope<
-      update_mirror_topic_status_response,
-      serde::version<0>,
-      serde::compat_version<0>> {
-    cluster_link::errc ec{cluster_link::errc::success};
-
-    friend bool operator==(
-      const update_mirror_topic_status_response&,
-      const update_mirror_topic_status_response&) = default;
-
-    auto serde_fields() { return std::tie(ec); }
-};
-
-struct batch_update_mirror_topic_status_request
-  : serde::envelope<
-      batch_update_mirror_topic_status_request,
-      serde::version<0>,
-      serde::compat_version<0>> {
-    ::cluster_link::model::id_t link_id;
-    ::cluster_link::model::batch_update_mirror_topic_status_cmd cmd;
-    model::timeout_clock::duration timeout{};
-
-    friend bool operator==(
-      const batch_update_mirror_topic_status_request&,
-      const batch_update_mirror_topic_status_request&) = default;
-
-    auto serde_fields() { return std::tie(link_id, cmd, timeout); }
-};
-
-struct batch_update_mirror_topic_status_response
-  : serde::envelope<
-      batch_update_mirror_topic_status_response,
-      serde::version<0>,
-      serde::compat_version<0>> {
-    cluster_link::errc ec{cluster_link::errc::success};
-
-    friend bool operator==(
-      const batch_update_mirror_topic_status_response&,
-      const batch_update_mirror_topic_status_response&) = default;
-
-    auto serde_fields() { return std::tie(ec); }
-};
-
-struct update_mirror_topic_properties_request
-  : serde::envelope<
-      update_mirror_topic_properties_request,
-      serde::version<0>,
-      serde::compat_version<0>> {
-    ::cluster_link::model::id_t link_id;
-    ::cluster_link::model::update_mirror_topic_properties_cmd cmd;
-    model::timeout_clock::duration timeout{};
-
-    friend bool operator==(
-      const update_mirror_topic_properties_request&,
-      const update_mirror_topic_properties_request&) = default;
-
-    auto serde_fields() { return std::tie(link_id, cmd, timeout); }
-};
-
-struct update_mirror_topic_properties_response
-  : serde::envelope<
-      update_mirror_topic_properties_response,
-      serde::version<0>,
-      serde::compat_version<0>> {
-    cluster_link::errc ec{cluster_link::errc::success};
-
-    friend bool operator==(
-      const update_mirror_topic_properties_response&,
-      const update_mirror_topic_properties_response&) = default;
-
-    auto serde_fields() { return std::tie(ec); }
-};
-
-struct update_cluster_link_configuration_request
-  : serde::envelope<
-      update_cluster_link_configuration_request,
-      serde::version<0>,
-      serde::compat_version<0>> {
-    ::cluster_link::model::id_t link_id;
-    ::cluster_link::model::update_cluster_link_configuration_cmd cmd;
-    model::timeout_clock::duration timeout{};
-
-    friend bool operator==(
-      const update_cluster_link_configuration_request&,
-      const update_cluster_link_configuration_request&) = default;
-
-    auto serde_fields() { return std::tie(link_id, cmd, timeout); }
-};
-
-struct update_cluster_link_configuration_response
-  : serde::envelope<
-      update_cluster_link_configuration_response,
-      serde::version<0>,
-      serde::compat_version<0>> {
-    cluster_link::errc ec{cluster_link::errc::success};
-
-    friend bool operator==(
-      const update_cluster_link_configuration_response&,
-      const update_cluster_link_configuration_response&) = default;
-
-    auto serde_fields() { return std::tie(ec); }
-};
-
-struct delete_mirror_topic_request
-  : serde::envelope<
-      delete_mirror_topic_request,
-      serde::version<0>,
-      serde::compat_version<0>> {
-    ::cluster_link::model::id_t link_id;
-    ::cluster_link::model::delete_mirror_topic_cmd cmd;
-    model::timeout_clock::duration timeout{};
-
-    friend bool operator==(
-      const delete_mirror_topic_request&,
-      const delete_mirror_topic_request&) = default;
-
-    auto serde_fields() { return std::tie(link_id, cmd, timeout); }
-};
-
-struct delete_mirror_topic_response
-  : serde::envelope<
-      delete_mirror_topic_response,
-      serde::version<0>,
-      serde::compat_version<0>> {
-    cluster_link::errc ec{cluster_link::errc::success};
-
-    friend bool operator==(
-      const delete_mirror_topic_response&,
-      const delete_mirror_topic_response&) = default;
-
-    auto serde_fields() { return std::tie(ec); }
-};
-
-// Request to get the current cluster epoch.
-struct get_current_cluster_epoch_request
-  : serde::envelope<
-      get_current_cluster_epoch_request,
-      serde::version<0>,
-      serde::compat_version<0>> {
-    model::timeout_clock::duration timeout{};
-
-    friend bool operator==(
-      const get_current_cluster_epoch_request&,
-      const get_current_cluster_epoch_request&) = default;
-
-    auto serde_fields() { return std::tie(timeout); }
-};
-
-// The reply to the get_current_cluster_epoch_request.
-struct get_current_cluster_epoch_response
-  : serde::envelope<
-      get_current_cluster_epoch_response,
-      serde::version<0>,
-      serde::compat_version<0>> {
-    errc ec{errc::success};
-    int64_t epoch{-1};
-
-    friend bool operator==(
-      const get_current_cluster_epoch_response&,
-      const get_current_cluster_epoch_response&) = default;
-
-    auto serde_fields() { return std::tie(ec, epoch); }
 };
 
 struct error_info {
